iniciando aba de iniciar conversa

+++ webchat/models.py
@@ -15,8 +15,7 @@ def get_profile_picture_path(instance, filename):
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    user_name = models.CharField(max_length=25,)
-    status = models.CharField(max_length=225, blank=True)
+    user_name = models.CharField(max_length=25)
     profile_picture = models.ImageField(upload_to=get_profile_picture_path, blank=True,
                                         null=True, default=DEFAULT_PROFILE_PICTURE)
     friends = models.ManyToManyField('self', symmetrical=True, blank=True)
@@ -33,8 +32,9 @@ class Profile(models.Model):
 
         if profile and profile.profile_picture and profile.profile_picture != self.profile_picture:
             try:
-                os.remove(os.path.join(settings.MEDIA_ROOT,
-                          str(profile.profile_picture)))
+                if str(profile.profile_picture) != DEFAULT_PROFILE_PICTURE:
+                    os.remove(os.path.join(settings.MEDIA_ROOT,
+                                           str(profile.profile_picture)))
             except FileNotFoundError:
                 pass
